DjangoBlog/requirement.txt
arrow==1.2.3
asgiref==3.5.2
backports.zoneinfo==0.2.1
binaryornot==0.4.4
certifi==2022.9.14
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
cookiecutter==2.1.1
Django==4.1.1
django-browser-reload==1.6.0
django-ckeditor==6.5.1
django-jazzmin==2.5.0
django-js-asset==2.0.0
django-tailwind==3.4.0
django-tinymce==3.5.0
idna==3.4
Jinja2==3.1.2
jinja2-time==0.2.0
MarkupSafe==2.1.1
python-dateutil==2.8.2
python-slugify==6.1.2
PyYAML==6.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
urllib3==1.26.12
whitenoise==6.2.0
